refactor: 좋아요 응답을 통일화

+++ src/main/java/org/ktc2/cokaen/wouldyouin/member/persist/LikeableMember.java
@@ -13,7 +13,7 @@ public interface LikeableMember {
 
     String getIntro();
 
-    Long getFollowers();
+    Integer getFollowers();
 
     String getHashtag();
